flatten components into add page
